add: Logger

+++ src/tasks/tasks.service.ts
@@ -37,7 +37,7 @@ export class TasksService {
       id: number,
       user: User,
       ): Promise<void> {
-    const result = await this.taskRepository.delete( id,);
+    const result = await this.taskRepository.delete( {id, userId: user.id});
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID "" not found`);
     }
